fix(Evaluator): fixed issue with recursion
The environment was modifier based on called function name. This was
a problem when a function was assigned to a value and then called from
this value. Then the function name in recursion would be the variable
name whence it should have been the original function's name. It has
been fixed by adding function's name in it's declaration and in the
environment.